Move test methods
